Refact : AccountService
###작업 내용###
- all or nothing에 맞게 구현
- AccountService 수정
- AccountHistoryService 삭제
